fixed makefile once more and fixed naming issues

+++ libft/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: jkauker <jkauker@student.42heilbrnn.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 00:27:24 by jonask            #+#    #+#             */
-/*   Updated: 2023/10/09 10:14:15 by jkauker          ###   ########.fr       */
+/*   Updated: 2023/10/09 11:40:33 by jkauker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
