+++ libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: jkauker <jkauker@student.42heilbrnn.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 01:37:45 by jonask            #+#    #+#             */
-/*   Updated: 2023/10/09 10:52:49 by jkauker          ###   ########.fr       */
+/*   Updated: 2023/10/09 11:43:21 by jkauker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (src[i] != 0 && i < size)
